Lots of Updates
Added a finished game popup. Hard to test since location can't be spoofed right now, but as far as I know, it works.
Updated all activities to use SoundPool. Looking into bug as to why onPause() causes it to not work.
Added code in MapsActivity for a finished game popup, as well as an mp3 file for sound.
We are not including sounds for when a landmark becomes claimable... it's way too annoying. :)